ml_v2/orchastrater/pyproject.toml
[project]
name = "orchastrater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "numpy>=2.4.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
